Added runtime monitoring mechanism to the Actor
The Actor reads job start time and duration from the file called
RaythenaTimeMonitor.txt which is created by the batch submission script
in the root run directory of the entire Raythena job. While Actor is
running, it monitors the time with regular intervals (5 sec hardwired
for now), and when it determines that it has 15min (hardwired again)
before the allocation expires, it creates and empty file named
pilot_kill_payload.

As the name suggests, when the pilot sees this file in its run
directory, it sends SIGKILL to AthenaMP, creates log tarball, reports
it back to the Actor in the final job update message, and stops.